refactor: split structural audit records
